Gather all init code of faraday to Presto::Client.faraday_client
Some were in `StatementClient#initialize` and some were in query.rb.
