+++ spec/statement_client_spec.rb
@@ -25,6 +25,10 @@ describe Presto::Client::StatementClient do
     }
   end
 
+  let :faraday do
+    Presto::Client.faraday_client(options)
+  end
+
   it "sets headers" do
     stub_request(:post, "localhost/v1/statement").
       with(body: query,
@@ -38,7 +42,6 @@ describe Presto::Client::StatementClient do
               "X-Presto-Session" => options[:properties].map {|k,v| "#{k}=#{v}"}.join("\r\nX-Presto-Session: ")
     }).to_return(body: response_json.to_json)
 
-    faraday = Faraday.new(url: "http://localhost")
     StatementClient.new(faraday, query, options)
   end
 
@@ -75,7 +78,6 @@ describe Presto::Client::StatementClient do
               "X-Presto-Session" => options[:properties].map {|k,v| "#{k}=#{v}"}.join("\r\nX-Presto-Session: ")
     }).to_return(body: lambda{|req|if retry_p; response_json.to_json; else; retry_p=true; raise Timeout::Error.new("execution expired"); end })
 
-    faraday = Faraday.new(url: "http://localhost")
     sc = StatementClient.new(faraday, query, options.merge(http_open_timeout: 1))
     sc.has_next?.should be_true
     sc.advance.should be_true
@@ -109,8 +111,8 @@ describe Presto::Client::StatementClient do
               "Accept" => "application/x-msgpack,application/json"
     }).to_return(body: lambda{|req|if retry_p; MessagePack.dump(response_json); else; retry_p=true; raise Timeout::Error.new("execution expired"); end }, headers: {"Content-Type" => "application/x-msgpack"})
 
-    faraday = Faraday.new(url: "http://localhost")
-    sc = StatementClient.new(faraday, query, options.merge(http_open_timeout: 1, enable_x_msgpack: "application/x-msgpack"))
+    options.merge!(http_open_timeout: 1, enable_x_msgpack: "application/x-msgpack")
+    sc = StatementClient.new(faraday, query, options)
     sc.has_next?.should be_true
     sc.advance.should be_true
     retry_p.should be_true
@@ -183,7 +185,7 @@ describe Presto::Client::StatementClient do
             basic_auth: [options[:user], password]
       ).to_return(body: response_json.to_json)
 
-      faraday = Query.__send__(:faraday_client, options.merge(ssl: { verify: true }, password: password))
+      options.merge!(ssl: { verify: true }, password: password)
       StatementClient.new(faraday, query, options)
     end
 
@@ -318,7 +320,6 @@ describe Presto::Client::StatementClient do
                  "X-Presto-Session" => options[:properties].map {|k,v| "#{k}=#{v}"}.join("\r\nX-Presto-Session: ")
              }).to_return(body: nested_json.to_json(:max_nesting => false))
 
-    faraday = Faraday.new(url: "http://localhost")
     StatementClient.new(faraday, query, options)
   end
 end
